shopsite/shopsite/requirements.txt
django==2.2.6
Pillow==6.2.0
WeasyPrint==0.42.3
braintree==3.57.1
celery==4.3.0
coreapi==2.3.3
django-parler==2.0
django-rosetta==0.8.1
djangorestframework==3.10.3
eventlet==0.25.1
flower==0.9.2
Jinja2==2.10.3
Pygments==2.4.2
amqp==2.5.1
django-redis==4.10.0
redis==3.3.10
hiredis==1.0.0
psycopg2==2.8.4